Refactor: move identical get_autoscaling_params from marathon_tools/kubernetes_tools to long_running_service_tools
